requirements.txt
asyncio
asyncpg
boto3==1.9.124
celery
discord.py
Django==2.1.7
django-autocomplete-light
django-celery-beat
django-crispy-forms
django-redis-cache
django-sendgrid-v5==0.8.0
django-storages==1.7.1
gunicorn
numpy==1.16.2
Pillow==5.4.1
pinax-messages
psycopg2-binary
pytz==2018.9
redis
requests==2.21.0
whitenoise==4.1.2